requirements.txt
azure-functions 
azure-storage-blob>=12.19.0